feat: refactor type handling in encode and decode implementations; add shared type utilities
