+++ messagepack-derive/src/lib.rs
@@ -5,6 +5,7 @@
 mod attrs;
 mod decode;
 mod encode;
+mod types;
 
 use proc_macro::TokenStream;
 use syn::{DeriveInput, parse_macro_input};
